fix: make OpenCL use 64-bit limbs
The refactoring that moved code from bellperson into this library lead
to a regression. Accidentally OpenCL was using 32-bit limbs, which
should lead to less performance than using 64-bit limbs.

The CUDA code path was always using the expected 32-bit limbs. Even
if it doesn't look like it from this change, it was the case as the
CUDA kernels are compiled at compile-time, the limb size is correctly
specified in build.rs.