feat: add a benchmark for multiexp (#23)
It is now possible to benchmark multiexp. The largest benchmark likely
needs more memory than the GPU has. This is done in order to make sure
that splitting the data into two runs also performs well.

With the default features, the benchmark is run on CUDA. To make sure
it is run with OpenCL only you can either set `EC_GPU_FRAMEWORK=opencl`
or disable the `cuda` feature via.

    cargo bench --no-default-features --features multiexp,opencl multiexp