feat: re-organize ec-gpu traits
The ec-gpu traits are re-organized.

BREAKING CHANGE: The `GpuEngine` trait is removed, as it was specific
to pairing friendly elliptic curves.

The `GpuField` trait gains a new method called `sub_field_name`, which
only matters for extension fields. It has a default implementation, so
existing implementations don't need to change.

A new trait called `GpuName` is introduced, which every `GpuField` needs
to implement. It contains the name of the object it's implemented for,
which is then used in the generated source code as identifier. As the
name shoule be unique, a macro called `name` is introduced. This way
implementations can have unique names without much hassle. Example:

    struct Fp;

    impl ec_gpu::GpuName for Fp {
        fn name() -> String {
            ec_gpu::name!()
        }
    }