fix: sending a result should be best effort
The threadpool is used for asynchronous tasks. Due to errors it can
happen that the result is no longer needed. Instead of panicking when
there's no receiving end anymore, only log a message.

Concretely this happens in `bellperson`, where we run the multiexp
operation in parallel on the GPU and the CPU. If there is an error
on the GPU part, we also discard the CPU part. We don't wait for the
CPU operations to be finished, but return early. Hence the receiving
will not be available anymore.