fix: improve DensityTracker::extend
`DensityTracker::extend` takes its `other` argument by value to avoid a memory
copy when the receiver's bitvec is empty (in most other cases, the `other`
bitvec is copied). This optimization is costly: the signature forces the caller
to relinquish allocated variants of `other` in all calls, including those where
the receiver is not empty.

Since the underlying `BitVec` already has a
[very efficient copy routine](https://docs.rs/bitvec/latest/bitvec/vec/struct.BitVec.html#method.copy_from_bitslice)
anyway, a better approach seems to be passing in the `other` argument by
reference.

BREAKING CHANGE: `DensityTracker::extend` now takes the `other` argument by
reference.